rename small_objects_counter -> objects_counter
