+++ objects_counter/consts.py
